tests/end2end: update navigation tests
