+++ tests/end2end/navigation/toc/input/document.sdoc
