+++ tests/end2end/navigation/viewtype_selector/expected_output/document.sdoc
