+++ tests/end2end/navigation/viewtype_selector/input/document.sdoc
