+++ tests/end2end/navigation/viewtype_selector/test_UC113_UI_viewtype_selector.py
@@ -30,7 +30,9 @@ class Test_UC113_UI_viewtype_selector(BaseCase):
             screen_document.assert_header_document_title("Document title")
             screen_document.assert_text("Hello world!")
 
-            viewtype_selector = ViewType_Selector(self)
+            viewtype_selector: ViewType_Selector = (
+                screen_document.get_viewtype_selector()
+            )
 
             # menu is on the page and closed
             viewtype_selector.assert_viewtype_handler()
